# use scan method for improve performance