before
```
Finished in 0.03372 seconds (files took 0.15419 seconds to load)
33 examples, 0 failures
```


after
```
Finished in 0.0204 seconds (files took 0.08934 seconds to load)
33 examples, 0 failures
```